T@x.tex: There is no char ^^99 (TM) in the lmtt fonts
File utf8.tex has lots of UTF-8 chars, and
I wanted to see them byte by byte using
the LY1 (texnansi) coding, which is complete from ^^80 to ^^ff.
The only problem is that the lmtt fonts do not have a char
for the Trade Mark (TM), which is in slot ^^99 in LY1.
So I have used another one, ^^7f (127), which is usually invalid.